+++ create.sql
@@ -5,12 +5,12 @@ SET time_zone = "+00:00";
 /*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
 /*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
 /*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
-/*!40101 SET NAMES utf8 */;
+/*!40101 SET NAMES utf8mb4 */;
 
 --
 -- Database: `mailscanner`
 --
-CREATE DATABASE IF NOT EXISTS `mailscanner` DEFAULT CHARACTER SET utf8 COLLATE utf8_unicode_ci;
+CREATE DATABASE IF NOT EXISTS `mailscanner` DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_520_ci;
 USE `mailscanner`;
 
 -- --------------------------------------------------------
@@ -22,11 +22,11 @@ USE `mailscanner`;
 CREATE TABLE IF NOT EXISTS `audit_log` (
   `id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT,
   `timestamp` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-  `user` varchar(191) COLLATE utf8_unicode_ci NOT NULL DEFAULT '',
-  `ip_address` varchar(45) COLLATE utf8_unicode_ci NOT NULL DEFAULT '',
-  `action` text COLLATE utf8_unicode_ci NOT NULL,
+  `user` varchar(191) COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT '',
+  `ip_address` varchar(45) COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT '',
+  `action` text COLLATE utf8mb4_unicode_520_ci NOT NULL,
   PRIMARY KEY (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -36,10 +36,10 @@ CREATE TABLE IF NOT EXISTS `audit_log` (
 
 CREATE TABLE IF NOT EXISTS `autorelease` (
   `id` bigint(20) NOT NULL AUTO_INCREMENT,
-  `msg_id` varchar(255) COLLATE utf8_unicode_ci NOT NULL,
-  `uid` varchar(255) COLLATE utf8_unicode_ci NOT NULL,
+  `msg_id` varchar(255) COLLATE utf8mb4_unicode_520_ci NOT NULL,
+  `uid` varchar(255) COLLATE utf8mb4_unicode_520_ci NOT NULL,
   PRIMARY KEY (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -49,12 +49,12 @@ CREATE TABLE IF NOT EXISTS `autorelease` (
 
 CREATE TABLE IF NOT EXISTS `blocklist` (
   `id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT,
-  `to_address` text COLLATE utf8_unicode_ci,
-  `to_domain` text COLLATE utf8_unicode_ci,
-  `from_address` text COLLATE utf8_unicode_ci,
+  `to_address` text COLLATE utf8mb4_unicode_520_ci,
+  `to_domain` text COLLATE utf8mb4_unicode_520_ci,
+  `from_address` text COLLATE utf8mb4_unicode_520_ci,
   PRIMARY KEY (`id`),
   UNIQUE KEY `blocklist_uniq` (`to_address`(100),`from_address`(100))
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -64,21 +64,21 @@ CREATE TABLE IF NOT EXISTS `blocklist` (
 
 CREATE TABLE IF NOT EXISTS `inq` (
   `inq_id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT,
-  `id` mediumtext COLLATE utf8_unicode_ci,
+  `id` mediumtext COLLATE utf8mb4_unicode_520_ci,
   `cdate` date DEFAULT NULL,
   `ctime` time DEFAULT NULL,
-  `from_address` mediumtext COLLATE utf8_unicode_ci,
-  `to_address` mediumtext COLLATE utf8_unicode_ci,
-  `subject` mediumtext COLLATE utf8_unicode_ci,
-  `message` mediumtext COLLATE utf8_unicode_ci,
-  `size` mediumtext COLLATE utf8_unicode_ci,
-  `priority` mediumtext COLLATE utf8_unicode_ci,
-  `attempts` mediumtext COLLATE utf8_unicode_ci,
-  `lastattempt` mediumtext COLLATE utf8_unicode_ci,
-  `hostname` mediumtext COLLATE utf8_unicode_ci,
+  `from_address` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `to_address` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `subject` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `message` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `size` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `priority` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `attempts` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `lastattempt` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `hostname` mediumtext COLLATE utf8mb4_unicode_520_ci,
   PRIMARY KEY (`inq_id`),
   KEY `inq_hostname` (`hostname`(50))
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -89,15 +89,15 @@ CREATE TABLE IF NOT EXISTS `inq` (
 CREATE TABLE IF NOT EXISTS `maillog` (
   `maillog_id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT,
   `timestamp` timestamp NULL,
-  `id` mediumtext COLLATE utf8_unicode_ci,
+  `id` mediumtext COLLATE utf8mb4_unicode_520_ci,
   `size` bigint(20) DEFAULT '0',
-  `from_address` mediumtext COLLATE utf8_unicode_ci,
-  `from_domain` mediumtext COLLATE utf8_unicode_ci,
-  `to_address` mediumtext COLLATE utf8_unicode_ci,
-  `to_domain` mediumtext COLLATE utf8_unicode_ci,
-  `subject` mediumtext COLLATE utf8_unicode_ci,
-  `clientip` mediumtext COLLATE utf8_unicode_ci,
-  `archive` mediumtext COLLATE utf8_unicode_ci,
+  `from_address` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `from_domain` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `to_address` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `to_domain` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `subject` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `clientip` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `archive` mediumtext COLLATE utf8mb4_unicode_520_ci,
   `isspam` tinyint(1) DEFAULT '0',
   `ishighspam` tinyint(1) DEFAULT '0',
   `issaspam` tinyint(1) DEFAULT '0',
@@ -107,26 +107,26 @@ CREATE TABLE IF NOT EXISTS `maillog` (
   `spamallowlisted` tinyint(1) DEFAULT '0',
   `spamblocklisted` tinyint(1) DEFAULT '0',
   `sascore` decimal(7,2) DEFAULT '0.00',
-  `spamreport` mediumtext COLLATE utf8_unicode_ci,
+  `spamreport` mediumtext COLLATE utf8mb4_unicode_520_ci,
   `virusinfected` tinyint(1) DEFAULT '0',
   `nameinfected` tinyint(2) DEFAULT '0',
   `otherinfected` tinyint(1) DEFAULT '0',
-  `report` mediumtext COLLATE utf8_unicode_ci,
+  `report` mediumtext COLLATE utf8mb4_unicode_520_ci,
   `ismcp` tinyint(1) DEFAULT '0',
   `ishighmcp` tinyint(1) DEFAULT '0',
   `issamcp` tinyint(1) DEFAULT '0',
   `mcpallowlisted` tinyint(1) DEFAULT '0',
   `mcpblocklisted` tinyint(1) DEFAULT '0',
   `mcpsascore` decimal(7,2) DEFAULT '0.00',
-  `mcpreport` mediumtext COLLATE utf8_unicode_ci,
-  `hostname` mediumtext COLLATE utf8_unicode_ci,
+  `mcpreport` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `hostname` mediumtext COLLATE utf8mb4_unicode_520_ci,
   `date` date DEFAULT NULL,
   `time` time DEFAULT NULL,
-  `headers` mediumtext COLLATE utf8_unicode_ci,
-  `messageid` mediumtext COLLATE utf8_unicode_ci DEFAULT NULL,
+  `headers` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `messageid` mediumtext COLLATE utf8mb4_unicode_520_ci DEFAULT NULL,
   `quarantined` tinyint(1) DEFAULT '0',
-  `rblspamreport` mediumtext COLLATE utf8_unicode_ci DEFAULT NULL,
-  `token` CHAR(64) COLLATE utf8_unicode_ci DEFAULT NULL,
+  `rblspamreport` mediumtext COLLATE utf8mb4_unicode_520_ci DEFAULT NULL,
+  `token` CHAR(64) COLLATE utf8mb4_unicode_520_ci DEFAULT NULL,
   `released` tinyint(1) DEFAULT '0',
   `salearn` tinyint(1) DEFAULT '0',
   `last_update` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
@@ -142,7 +142,7 @@ CREATE TABLE IF NOT EXISTS `maillog` (
   KEY `maillog_quarantined` (`quarantined`),
   KEY `timestamp_idx` (`timestamp`)
   /*!50604 , FULLTEXT KEY `subject_idx` (`subject`) */
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -151,10 +151,10 @@ CREATE TABLE IF NOT EXISTS `maillog` (
 --
 
 CREATE TABLE IF NOT EXISTS `mcp_rules` (
-  `rule` varchar(191) COLLATE utf8_unicode_ci NOT NULL DEFAULT '',
-  `rule_desc` varchar(512) COLLATE utf8_unicode_ci NOT NULL DEFAULT '',
+  `rule` varchar(191) COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT '',
+  `rule_desc` varchar(512) COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT '',
   PRIMARY KEY (`rule`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -165,20 +165,20 @@ CREATE TABLE IF NOT EXISTS `mcp_rules` (
 CREATE TABLE IF NOT EXISTS `mtalog` (
   `mtalog_id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT,
   `timestamp` datetime DEFAULT NULL,
-  `host` mediumtext COLLATE utf8_unicode_ci,
-  `type` mediumtext COLLATE utf8_unicode_ci,
-  `msg_id` varchar(20) COLLATE utf8_unicode_ci DEFAULT NULL,
-  `relay` mediumtext COLLATE utf8_unicode_ci,
-  `dsn` mediumtext COLLATE utf8_unicode_ci,
-  `status` mediumtext COLLATE utf8_unicode_ci,
+  `host` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `type` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `msg_id` varchar(20) COLLATE utf8mb4_unicode_520_ci DEFAULT NULL,
+  `relay` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `dsn` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `status` mediumtext COLLATE utf8mb4_unicode_520_ci,
   `delay` time DEFAULT NULL,
-  `to_address` mediumtext COLLATE utf8_unicode_ci DEFAULT NULL,
+  `to_address` mediumtext COLLATE utf8mb4_unicode_520_ci DEFAULT NULL,
   PRIMARY KEY (`mtalog_id`),
   UNIQUE KEY `mtalog_uniq` (`timestamp`,`host`(10),`type`(10),`msg_id`,`relay`(20),`to_address`(64)),
   KEY `mtalog_timestamp` (`timestamp`),
   KEY `mtalog_type` (`type`(10)),
   KEY `msg_id` (`msg_id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -191,7 +191,7 @@ CREATE TABLE IF NOT EXISTS `mtalog_ids` (
   `smtp_id` varchar(20) CHARACTER SET ascii DEFAULT NULL,
   UNIQUE KEY `mtalog_ids_idx` (`smtpd_id`,`smtp_id`),
   KEY `smtpd_id` (`smtpd_id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -201,21 +201,21 @@ CREATE TABLE IF NOT EXISTS `mtalog_ids` (
 
 CREATE TABLE IF NOT EXISTS `outq` (
   `outq_id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT,
-  `id` mediumtext COLLATE utf8_unicode_ci,
+  `id` mediumtext COLLATE utf8mb4_unicode_520_ci,
   `cdate` date DEFAULT NULL,
   `ctime` time DEFAULT NULL,
-  `from_address` mediumtext COLLATE utf8_unicode_ci,
-  `to_address` mediumtext COLLATE utf8_unicode_ci,
-  `subject` mediumtext COLLATE utf8_unicode_ci,
-  `message` mediumtext COLLATE utf8_unicode_ci,
-  `size` mediumtext COLLATE utf8_unicode_ci,
-  `priority` mediumtext COLLATE utf8_unicode_ci,
-  `attempts` mediumtext COLLATE utf8_unicode_ci,
-  `lastattempt` mediumtext COLLATE utf8_unicode_ci,
-  `hostname` mediumtext COLLATE utf8_unicode_ci,
+  `from_address` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `to_address` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `subject` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `message` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `size` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `priority` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `attempts` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `lastattempt` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `hostname` mediumtext COLLATE utf8mb4_unicode_520_ci,
   PRIMARY KEY (`outq_id`),
   KEY `outq_hostname` (`hostname`(50))
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -225,14 +225,14 @@ CREATE TABLE IF NOT EXISTS `outq` (
 
 CREATE TABLE IF NOT EXISTS `saved_filters` (
   `id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT,
-  `name` mediumtext COLLATE utf8_unicode_ci NOT NULL,
-  `col` mediumtext COLLATE utf8_unicode_ci NOT NULL,
-  `operator` mediumtext COLLATE utf8_unicode_ci NOT NULL,
-  `value` mediumtext COLLATE utf8_unicode_ci NOT NULL,
-  `username` mediumtext COLLATE utf8_unicode_ci NOT NULL,
+  `name` mediumtext COLLATE utf8mb4_unicode_520_ci NOT NULL,
+  `col` mediumtext COLLATE utf8mb4_unicode_520_ci NOT NULL,
+  `operator` mediumtext COLLATE utf8mb4_unicode_520_ci NOT NULL,
+  `value` mediumtext COLLATE utf8mb4_unicode_520_ci NOT NULL,
+  `username` mediumtext COLLATE utf8mb4_unicode_520_ci NOT NULL,
   PRIMARY KEY (`id`),
   UNIQUE KEY `unique_filters` (`name`(20),`col`(20),`operator`(20),`value`(20),`username`(20))
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -241,10 +241,10 @@ CREATE TABLE IF NOT EXISTS `saved_filters` (
 --
 
 CREATE TABLE IF NOT EXISTS `sa_rules` (
-  `rule` varchar(191) COLLATE utf8_unicode_ci NOT NULL DEFAULT '',
-  `rule_desc` varchar(512) COLLATE utf8_unicode_ci NOT NULL DEFAULT '',
+  `rule` varchar(191) COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT '',
+  `rule_desc` varchar(512) COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT '',
   PRIMARY KEY (`rule`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -254,23 +254,23 @@ CREATE TABLE IF NOT EXISTS `sa_rules` (
 
 CREATE TABLE IF NOT EXISTS `users` (
   `id` BIGINT NOT NULL AUTO_INCREMENT UNIQUE KEY,
-  `username` varchar(191) COLLATE utf8_unicode_ci NOT NULL DEFAULT '',
-  `password` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
-  `fullname` varchar(255) COLLATE utf8_unicode_ci NOT NULL DEFAULT '',
-  `type` enum('A','D','U','R','H') COLLATE utf8_unicode_ci NOT NULL DEFAULT 'U',
+  `username` varchar(191) COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT '',
+  `password` varchar(255) COLLATE utf8mb4_unicode_520_ci DEFAULT NULL,
+  `fullname` varchar(255) COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT '',
+  `type` enum('A','D','U','R','H') COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT 'U',
   `quarantine_report` tinyint(1) DEFAULT '0',
   `spamscore` float DEFAULT '0',
   `highspamscore` float DEFAULT '0',
   `noscan` tinyint(1) DEFAULT '0',
-  `quarantine_rcpt` varchar(60) COLLATE utf8_unicode_ci DEFAULT NULL,
-  `resetid` varchar(32) COLLATE utf8_unicode_ci DEFAULT NULL,
-  `resetexpire` bigint(20) COLLATE utf8_unicode_ci DEFAULT NULL,
-  `lastreset` bigint(20) COLLATE utf8_unicode_ci DEFAULT NULL,
-  `login_expiry` bigint(20) COLLATE utf8_unicode_ci DEFAULT '-1',
-  `last_login` bigint(20) COLLATE utf8_unicode_ci DEFAULT '-1',
-  `login_timeout` smallint(5) COLLATE utf8_unicode_ci DEFAULT '-1',
+  `quarantine_rcpt` varchar(60) COLLATE utf8mb4_unicode_520_ci DEFAULT NULL,
+  `resetid` varchar(32) COLLATE utf8mb4_unicode_520_ci DEFAULT NULL,
+  `resetexpire` bigint(20) COLLATE utf8mb4_unicode_520_ci DEFAULT NULL,
+  `lastreset` bigint(20) COLLATE utf8mb4_unicode_520_ci DEFAULT NULL,
+  `login_expiry` bigint(20) COLLATE utf8mb4_unicode_520_ci DEFAULT '-1',
+  `last_login` bigint(20) COLLATE utf8mb4_unicode_520_ci DEFAULT '-1',
+  `login_timeout` smallint(5) COLLATE utf8mb4_unicode_520_ci DEFAULT '-1',
   PRIMARY KEY (`username`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -280,13 +280,13 @@ CREATE TABLE IF NOT EXISTS `users` (
 
 CREATE TABLE IF NOT EXISTS `user_filters` (
   `id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT,
-  `username` varchar(191) COLLATE utf8_unicode_ci NOT NULL DEFAULT '',
-  `filter` mediumtext COLLATE utf8_unicode_ci,
-  `verify_key` varchar(32) COLLATE utf8_unicode_ci NOT NULL DEFAULT '',
-  `active` enum('N','Y') COLLATE utf8_unicode_ci DEFAULT 'N',
+  `username` varchar(191) COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT '',
+  `filter` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `verify_key` varchar(32) COLLATE utf8mb4_unicode_520_ci NOT NULL DEFAULT '',
+  `active` enum('N','Y') COLLATE utf8mb4_unicode_520_ci DEFAULT 'N',
   PRIMARY KEY (`id`),
   KEY `user_filters_username_idx` (`username`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------
 
@@ -296,12 +296,12 @@ CREATE TABLE IF NOT EXISTS `user_filters` (
 
 CREATE TABLE IF NOT EXISTS `allowlist` (
   `id` bigint(11) UNSIGNED NOT NULL AUTO_INCREMENT,
-  `to_address` mediumtext COLLATE utf8_unicode_ci,
-  `to_domain` mediumtext COLLATE utf8_unicode_ci,
-  `from_address` mediumtext COLLATE utf8_unicode_ci,
+  `to_address` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `to_domain` mediumtext COLLATE utf8mb4_unicode_520_ci,
+  `from_address` mediumtext COLLATE utf8mb4_unicode_520_ci,
   PRIMARY KEY (`id`),
   UNIQUE KEY `allowlist_uniq` (`to_address`(100),`from_address`(100))
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 -- --------------------------------------------------------